Move `GoogleGroup` test env classes to upper-level package
